server/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
mysqlclient==1.4.1
pytz==2018.9
django-cors-headers==2.4.0